Update / Reformat - Implementing a switch based sectionControl
Inputs by FantasyTeddy:
Do not define variables at the top, but where they are actually being used. - Done

Could we maybe introduce 2 new methods: one called something like HandleButtonHardware with the content of this if block and one like HandleSwitchHardware with the content of the else if block? - Done

I think when dealing with bit-operations, it is best to use bit shift operations (like <<) instead of "proper" mathematical operations.
Maybe we could also think about adding some helper functions that would simplify such bit checking...? - Done

Only use double // for comments (triple /// are used for documentation comments). - Done

Does this need to be an else if or could it just be else, because you are just checking the inverse of the if? - Done

Here too: maybe add a method HandleButtonZones for the if block and a HandleSwitchZones for the else if block. - Done, solved with point 2