Close confirmation, Advanced Settings Panel, added StartMinimized option  (#883)
* Adds start minimized option

Adds a "Start Minimized" option to the AgIO form.

This allows the user to configure the application to start in a minimized state, saving space on the taskbar.
The option is persisted to the application settings.

* Adds confirmation dialog on exit

Adds a confirmation dialog when the user attempts to exit the application.

This prevents accidental closure of the application by requiring the user to explicitly confirm their intention to exit.
Also moves the start minimized checkbox and the minimize button to the top right of the window.

* Adds advanced settings form

Adds a new form for advanced settings, including options for
starting minimized and auto-running GPS-Out.
Removes the checkboxes controlling these settings from the main form and
places them in the new advanced settings form.
This commit also adds a minimize icon.